Three dimensional Hilbert curve

Three dimensional Hilbert curve

hilbert_3d(level = 2L)

Arguments

level

The level of the curve.

Value

A three-column matrix of the coordinates of the Hilbert curve segments.

Details

There are many forms of 3D Hilbert curve. Here we only implement one specific form of it.

See also

Examples

pos = hilbert_3d(2)
if(require(rgl) && interactive()) {
    plot3d(pos, type = "l", lwd = 4, col = 2)
}
#> Loading required package: rgl