Rules

A list of pre-defined expansion rules for different curves.

Usage

SFC_RULES_2x2

SFC_RULES_3x3_PEANO

SFC_RULES_3x3_MEANDER

SFC_RULES_3x3_COMBINED

SFC_RULES_4x4_MEANDER_1

SFC_RULES_4x4_MEANDER_2

Format

An object of class sfc_rules of length 1.

An object of class sfc_rules of length 1.

An object of class sfc_rules of length 1.

An object of class sfc_rules of length 1.

An object of class sfc_rules of length 1.

An object of class sfc_rules of length 1.

Details

SFC_RULES_3x3_PEANO, SFC_RULES_3x3_MEANDER and SFC_RULES_3x3_COMBINED, SFC_RULES_MEANDER_4x4_1, SFC_RULES_MEANDER_4x4_2 also contain the "flipped" expansion rules.

SFC_RULES_3x3_COMBINED is a combination of SFC_RULES_3x3_PEANO and SFC_RULES_3x3_MEANDER where in SFC_RULES_3x3_PEANO, J is replaced by its original pattern I.

SFC_RULES_4x4_MEANDER_1 and SFC_RULES_4x4_MEANDER_2 are extension rules of Meander curves (3x3) on the 4x4 curve. It is only for the demonstration purpose, thus only I/R/L are supported.

Examples

SFC_RULES_2x2
#> Name: 2x2
#> I | I_1 = R(0)L(270)L(0)R(90)    corner = (1, 2) 
#>     I_2 = L(0)R(90)R(0)L(270)    corner = (2, 1) 
#> R | R_1 = I(0)R(0)R(270)L(180)   corner = (1, 2) 
#>     R_2 = L(0)R(90)R(0)I(270)    corner = (2, 1) 
#> L | L_1 = R(0)L(270)L(0)I(90)    corner = (1, 2) 
#>     L_2 = I(0)L(0)L(90)R(180)    corner = (2, 1) 
#> U | U_1 = I(0)R(0)R(270)I(180)   corner = (1, 2) 
#>     U_2 = I(0)L(0)L(90)I(180)    corner = (2, 1) 
#> B | B_1 = L(180)L(270)L(0)R(90)  corner = (1, 2) 
#>     B_2 = R(270)R(180)R(90)I(0)  corner = (1, 2) 
#> D | D_1 = L(90)L(180)L(270)I(0)  corner = (2, 1) 
#>     D_2 = R(180)R(90)R(0)L(270)  corner = (2, 1) 
#> P | P_1 = R(0)L(270)L(0)L(90)    corner = (1, 2) 
#>     P_2 = I(0)R(0)R(270)R(180)   corner = (1, 2) 
#> Q | Q_1 = I(0)L(0)L(90)L(180)    corner = (2, 1) 
#>     Q_2 = L(0)R(90)R(0)R(270)    corner = (2, 1) 
#> C | C_1 = L(270)L(0)L(90)L(180)  corner = (2, 1) 
#>     C_2 = R(90)R(0)R(270)R(180)  corner = (1, 2) 
SFC_RULES_3x3_PEANO
#> Name: 3x3 Peano
#> I | I_1 = I(0)J(0)R(0)R(270)I(180)L(180)L(270)J(0)I(0)  corner = (1, 1) 
#> J | J_1 = J(0)I(0)L(0)L(90)J(180)R(180)R(90)I(0)J(0)    corner = (2, 2) 
#> R | R_1 = I(0)J(0)R(0)R(270)I(180)L(180)L(270)J(0)R(0)  corner = (1, 1) 
#> L | L_1 = J(0)I(0)L(0)L(90)J(180)R(180)R(90)I(0)L(0)    corner = (2, 2) 
#> 
#> Flipped:
#> I | I_1 = R(0)I(270)L(270)L(0)J(90)R(90)R(0)I(270)L(270)  corner = (1, 1) 
#> J | J_1 = L(0)J(90)R(90)R(0)I(270)L(270)L(0)J(90)R(90)    corner = (2, 2) 
#> R | R_1 = R(0)I(270)L(270)L(0)J(90)R(90)R(0)I(270)J(270)  corner = (1, 1) 
#> L | L_1 = L(0)J(90)R(90)R(0)I(270)L(270)L(0)J(90)I(90)    corner = (2, 2) 
SFC_RULES_3x3_MEANDER
#> Name: 3x3 Meander
#> I | I_1 = R(0)I(270)L(270)I(0)L(0)L(90)R(180)R(90)I(0)      corner = (1, 2) 
#>     I_2 = L(0)I(90)R(90)I(0)R(0)R(270)L(180)L(270)I(0)      corner = (2, 1) 
#> R | R_1 = I(0)I(0)R(0)I(270)R(270)R(180)L(90)L(180)I(270)   corner = (1, 2) 
#>     R_2 = L(0)I(90)R(90)I(0)R(0)R(270)L(180)L(270)R(0)      corner = (2, 1) 
#> L | L_1 = R(0)I(270)L(270)I(0)L(0)L(90)R(180)R(90)L(0)      corner = (1, 2) 
#>     L_2 = I(0)I(0)L(0)I(90)L(90)L(180)R(270)R(180)I(90)     corner = (2, 1) 
#> U | U_1 = I(0)I(0)R(0)I(270)R(270)R(180)L(90)L(180)R(270)   corner = (1, 2) 
#>     U_2 = I(0)I(0)L(0)I(90)L(90)L(180)R(270)R(180)L(90)     corner = (2, 1) 
#> B | B_1 = L(180)I(270)L(270)I(0)L(0)L(90)R(180)R(90)I(0)    corner = (1, 2) 
#>     B_2 = R(270)I(180)R(180)I(90)R(90)R(0)L(270)L(0)R(90)   corner = (1, 2) 
#> D | D_1 = L(90)I(180)L(180)I(270)L(270)L(0)R(90)R(0)L(270)  corner = (2, 1) 
#>     D_2 = R(180)I(90)R(90)I(0)R(0)R(270)L(180)L(270)I(0)    corner = (2, 1) 
#> P | P_1 = I(0)I(0)R(0)I(270)R(270)R(180)L(90)L(180)L(270)   corner = (1, 2) 
#>     P_2 = R(0)I(270)L(270)I(0)L(0)L(90)R(180)R(90)R(0)      corner = (1, 2) 
#> Q | Q_1 = I(0)I(0)L(0)I(90)L(90)L(180)R(270)R(180)R(90)     corner = (2, 1) 
#>     Q_2 = L(0)I(90)R(90)I(0)R(0)R(270)L(180)L(270)L(0)      corner = (2, 1) 
#> C | C_1 = L(0)I(90)L(90)I(180)L(180)L(270)R(0)R(270)R(180)  corner = (1, 2) 
#>     C_2 = R(0)I(270)R(270)I(180)R(180)R(90)L(0)L(90)L(180)  corner = (2, 1) 
#> 
#> Flipped:
#> I | I_1 = I(0)R(0)R(270)L(180)L(270)I(0)L(0)I(90)R(90)        corner = (1, 2) 
#>     I_2 = I(0)L(0)L(90)R(180)R(90)I(0)R(0)I(270)L(270)        corner = (2, 1) 
#> R | R_1 = R(0)L(270)L(0)R(90)R(0)I(270)R(270)I(180)L(180)     corner = (1, 2) 
#>     R_2 = I(0)L(0)L(90)R(180)R(90)I(0)R(0)I(270)I(270)        corner = (2, 1) 
#> L | L_1 = I(0)R(0)R(270)L(180)L(270)I(0)L(0)I(90)I(90)        corner = (1, 2) 
#>     L_2 = L(0)R(90)R(0)L(270)L(0)I(90)L(90)I(180)R(180)       corner = (2, 1) 
#> U | U_1 = R(0)L(270)L(0)R(90)R(0)I(270)R(270)I(180)I(180)     corner = (1, 2) 
#>     U_2 = L(0)R(90)R(0)L(270)L(0)I(90)L(90)I(180)I(180)       corner = (2, 1) 
#> B | B_1 = R(90)R(0)R(270)L(180)L(270)I(0)L(0)I(90)R(90)       corner = (1, 2) 
#>     B_2 = L(0)L(90)L(180)R(270)R(180)I(90)R(90)I(0)I(0)       corner = (1, 2) 
#> D | D_1 = R(0)R(270)R(180)L(90)L(180)I(270)L(270)I(0)I(0)     corner = (2, 1) 
#>     D_2 = L(270)L(0)L(90)R(180)R(90)I(0)R(0)I(270)L(270)      corner = (2, 1) 
#> P | P_1 = R(0)L(270)L(0)R(90)R(0)I(270)R(270)I(180)R(180)     corner = (1, 2) 
#>     P_2 = I(0)R(0)R(270)L(180)L(270)I(0)L(0)I(90)L(90)        corner = (1, 2) 
#> Q | Q_1 = L(0)R(90)R(0)L(270)L(0)I(90)L(90)I(180)L(180)       corner = (2, 1) 
#>     Q_2 = I(0)L(0)L(90)R(180)R(90)I(0)R(0)I(270)R(270)        corner = (2, 1) 
#> C | C_1 = R(270)R(180)R(90)L(0)L(90)I(180)L(180)I(270)L(270)  corner = (1, 2) 
#>     C_2 = L(90)L(180)L(270)R(0)R(270)I(180)R(180)I(90)R(90)   corner = (2, 1) 
SFC_RULES_3x3_COMBINED
#> Name: 3x3 combined
#> I | I_1 = R(0)I(270)L(270)I(0)L(0)L(90)R(180)R(90)I(0)      corner = (1, 2) 
#>     I_2 = L(0)I(90)R(90)I(0)R(0)R(270)L(180)L(270)I(0)      corner = (2, 1) 
#>     I_3 = I(0)I(0)R(0)R(270)I(180)L(180)L(270)I(0)I(0)      corner = (1, 1) 
#>     I_4 = I(0)I(0)L(0)L(90)I(180)R(180)R(90)I(0)I(0)        corner = (2, 2) 
#> R | R_1 = I(0)I(0)R(0)I(270)R(270)R(180)L(90)L(180)I(270)   corner = (1, 2) 
#>     R_2 = L(0)I(90)R(90)I(0)R(0)R(270)L(180)L(270)R(0)      corner = (2, 1) 
#>     R_3 = I(0)I(0)R(0)R(270)I(180)L(180)L(270)I(0)R(0)      corner = (1, 1) 
#> L | L_1 = R(0)I(270)L(270)I(0)L(0)L(90)R(180)R(90)L(0)      corner = (1, 2) 
#>     L_2 = I(0)I(0)L(0)I(90)L(90)L(180)R(270)R(180)I(90)     corner = (2, 1) 
#>     L_3 = I(0)I(0)L(0)L(90)I(180)R(180)R(90)I(0)L(0)        corner = (2, 2) 
#> U | U_1 = I(0)I(0)R(0)I(270)R(270)R(180)L(90)L(180)R(270)   corner = (1, 2) 
#>     U_2 = I(0)I(0)L(0)I(90)L(90)L(180)R(270)R(180)L(90)     corner = (2, 1) 
#> B | B_1 = L(180)I(270)L(270)I(0)L(0)L(90)R(180)R(90)I(0)    corner = (1, 2) 
#>     B_2 = R(270)I(180)R(180)I(90)R(90)R(0)L(270)L(0)R(90)   corner = (1, 2) 
#>     B_3 = L(270)I(0)L(0)L(90)I(180)R(180)R(90)I(0)I(0)      corner = (2, 2) 
#> D | D_1 = L(90)I(180)L(180)I(270)L(270)L(0)R(90)R(0)L(270)  corner = (2, 1) 
#>     D_2 = R(180)I(90)R(90)I(0)R(0)R(270)L(180)L(270)I(0)    corner = (2, 1) 
#>     D_3 = R(90)I(0)R(0)R(270)I(180)L(180)L(270)I(0)I(0)     corner = (1, 1) 
#> P | P_1 = I(0)I(0)R(0)I(270)R(270)R(180)L(90)L(180)L(270)   corner = (1, 2) 
#>     P_2 = R(0)I(270)L(270)I(0)L(0)L(90)R(180)R(90)R(0)      corner = (1, 2) 
#>     P_3 = I(0)I(0)R(0)R(270)I(180)L(180)L(270)I(0)L(0)      corner = (1, 1) 
#> Q | Q_1 = I(0)I(0)L(0)I(90)L(90)L(180)R(270)R(180)R(90)     corner = (2, 1) 
#>     Q_2 = L(0)I(90)R(90)I(0)R(0)R(270)L(180)L(270)L(0)      corner = (2, 1) 
#>     Q_3 = I(0)I(0)L(0)L(90)I(180)R(180)R(90)I(0)R(0)        corner = (2, 2) 
#> C | C_1 = L(0)I(90)L(90)I(180)L(180)L(270)R(0)R(270)R(180)  corner = (1, 2) 
#>     C_2 = R(0)I(270)R(270)I(180)R(180)R(90)L(0)L(90)L(180)  corner = (2, 1) 
#>     C_3 = R(90)I(0)R(0)R(270)I(180)L(180)L(270)I(0)L(0)     corner = (1, 1) 
#>     C_4 = L(270)I(0)L(0)L(90)I(180)R(180)R(90)I(0)R(0)      corner = (2, 2) 
#> 
#> Flipped:
#> I | I_1 = I(0)R(0)R(270)L(180)L(270)I(0)L(0)I(90)R(90)        corner = (1, 2) 
#>     I_2 = I(0)L(0)L(90)R(180)R(90)I(0)R(0)I(270)L(270)        corner = (2, 1) 
#>     I_3 = R(0)I(270)L(270)L(0)I(90)R(90)R(0)I(270)L(270)      corner = (1, 1) 
#>     I_4 = L(0)I(90)R(90)R(0)I(270)L(270)L(0)I(90)R(90)        corner = (2, 2) 
#> R | R_1 = R(0)L(270)L(0)R(90)R(0)I(270)R(270)I(180)L(180)     corner = (1, 2) 
#>     R_2 = I(0)L(0)L(90)R(180)R(90)I(0)R(0)I(270)I(270)        corner = (2, 1) 
#>     R_3 = R(0)I(270)L(270)L(0)I(90)R(90)R(0)I(270)I(270)      corner = (1, 1) 
#> L | L_1 = I(0)R(0)R(270)L(180)L(270)I(0)L(0)I(90)I(90)        corner = (1, 2) 
#>     L_2 = L(0)R(90)R(0)L(270)L(0)I(90)L(90)I(180)R(180)       corner = (2, 1) 
#>     L_3 = L(0)I(90)R(90)R(0)I(270)L(270)L(0)I(90)I(90)        corner = (2, 2) 
#> U | U_1 = R(0)L(270)L(0)R(90)R(0)I(270)R(270)I(180)I(180)     corner = (1, 2) 
#>     U_2 = L(0)R(90)R(0)L(270)L(0)I(90)L(90)I(180)I(180)       corner = (2, 1) 
#> B | B_1 = R(90)R(0)R(270)L(180)L(270)I(0)L(0)I(90)R(90)       corner = (1, 2) 
#>     B_2 = L(0)L(90)L(180)R(270)R(180)I(90)R(90)I(0)I(0)       corner = (1, 2) 
#>     B_3 = U(270)I(90)R(90)R(0)I(270)L(270)L(0)I(90)R(90)      corner = (2, 2) 
#> D | D_1 = R(0)R(270)R(180)L(90)L(180)I(270)L(270)I(0)I(0)     corner = (2, 1) 
#>     D_2 = L(270)L(0)L(90)R(180)R(90)I(0)R(0)I(270)L(270)      corner = (2, 1) 
#>     D_3 = U(90)I(270)L(270)L(0)I(90)R(90)R(0)I(270)L(270)     corner = (1, 1) 
#> P | P_1 = R(0)L(270)L(0)R(90)R(0)I(270)R(270)I(180)R(180)     corner = (1, 2) 
#>     P_2 = I(0)R(0)R(270)L(180)L(270)I(0)L(0)I(90)L(90)        corner = (1, 2) 
#>     P_3 = R(0)I(270)L(270)L(0)I(90)R(90)R(0)I(270)U(270)      corner = (1, 1) 
#> Q | Q_1 = L(0)R(90)R(0)L(270)L(0)I(90)L(90)I(180)L(180)       corner = (2, 1) 
#>     Q_2 = I(0)L(0)L(90)R(180)R(90)I(0)R(0)I(270)R(270)        corner = (2, 1) 
#>     Q_3 = L(0)I(90)R(90)R(0)I(270)L(270)L(0)I(90)U(90)        corner = (2, 2) 
#> C | C_1 = R(270)R(180)R(90)L(0)L(90)I(180)L(180)I(270)L(270)  corner = (1, 2) 
#>     C_2 = L(90)L(180)L(270)R(0)R(270)I(180)R(180)I(90)R(90)   corner = (2, 1) 
#>     C_3 = U(90)I(270)L(270)L(0)I(90)R(90)R(0)I(270)U(270)     corner = (1, 1) 
#>     C_4 = U(270)I(90)R(90)R(0)I(270)L(270)L(0)I(90)U(90)      corner = (2, 2) 
SFC_RULES_4x4_MEANDER_1
#> Name: 4x4 Meander type 1
#> I | I_1 = I(0)I(0)R(0)I(270)R(270)R(180)L(90)L(180)I(270)L(270)I(0)I(0)L(0)I(90)I(90)R(90)      corner = (1, 2) 
#>     I_2 = I(0)I(0)L(0)I(90)L(90)L(180)R(270)R(180)I(90)R(90)I(0)I(0)R(0)I(270)I(270)L(270)      corner = (2, 1) 
#> R | R_1 = R(0)I(270)L(270)I(0)L(0)L(90)R(180)R(90)I(0)R(0)I(270)I(270)R(270)I(180)I(180)L(180)  corner = (1, 2) 
#>     R_2 = I(0)I(0)L(0)I(90)L(90)L(180)R(270)R(180)I(90)R(90)I(0)I(0)R(0)I(270)I(270)I(270)      corner = (2, 1) 
#> L | L_1 = I(0)I(0)R(0)I(270)R(270)R(180)L(90)L(180)I(270)L(270)I(0)I(0)L(0)I(90)I(90)I(90)      corner = (1, 2) 
#>     L_2 = L(0)I(90)R(90)I(0)R(0)R(270)L(180)L(270)I(0)L(0)I(90)I(90)L(90)I(180)I(180)R(180)     corner = (2, 1) 
#> 
#> Flipped:
#> I | I_1 = R(0)I(270)I(270)L(270)I(0)I(0)L(0)I(90)L(90)L(180)R(270)R(180)I(90)R(90)I(0)I(0)      corner = (1, 2) 
#>     I_2 = L(0)I(90)I(90)R(90)I(0)I(0)R(0)I(270)R(270)R(180)L(90)L(180)I(270)L(270)I(0)I(0)      corner = (2, 1) 
#> R | R_1 = I(0)I(0)I(0)R(0)I(270)I(270)R(270)I(180)R(180)R(90)L(0)L(90)I(180)L(180)I(270)I(270)  corner = (1, 2) 
#>     R_2 = L(0)I(90)I(90)R(90)I(0)I(0)R(0)I(270)R(270)R(180)L(90)L(180)I(270)L(270)I(0)R(0)      corner = (2, 1) 
#> L | L_1 = R(0)I(270)I(270)L(270)I(0)I(0)L(0)I(90)L(90)L(180)R(270)R(180)I(90)R(90)I(0)L(0)      corner = (1, 2) 
#>     L_2 = I(0)I(0)I(0)L(0)I(90)I(90)L(90)I(180)L(180)L(270)R(0)R(270)I(180)R(180)I(90)I(90)     corner = (2, 1) 
SFC_RULES_4x4_MEANDER_2
#> Name: 4x4 Meander type 2
#> I | I_1 = R(0)I(270)I(270)L(270)I(0)I(0)L(0)L(90)I(180)R(180)I(90)R(90)R(0)L(270)L(0)R(90)      corner = (1, 2) 
#>     I_2 = L(0)I(90)I(90)R(90)I(0)I(0)R(0)R(270)I(180)L(180)I(270)L(270)L(0)R(90)R(0)L(270)      corner = (2, 1) 
#> R | R_1 = I(0)I(0)I(0)R(0)I(270)I(270)R(270)R(180)I(90)L(90)I(180)L(180)L(270)R(0)R(270)L(180)  corner = (1, 2) 
#>     R_2 = L(0)I(90)I(90)R(90)I(0)I(0)R(0)R(270)I(180)L(180)I(270)L(270)L(0)R(90)R(0)I(270)      corner = (2, 1) 
#> L | L_1 = R(0)I(270)I(270)L(270)I(0)I(0)L(0)L(90)I(180)R(180)I(90)R(90)R(0)L(270)L(0)I(90)      corner = (1, 2) 
#>     L_2 = I(0)I(0)I(0)L(0)I(90)I(90)L(90)L(180)I(270)R(270)I(180)R(180)R(90)L(0)L(90)R(180)     corner = (2, 1) 
#> 
#> Flipped:
#> I | I_1 = R(0)L(270)L(0)R(90)R(0)I(270)R(270)I(180)L(180)L(270)I(0)I(0)L(0)I(90)I(90)R(90)      corner = (1, 2) 
#>     I_2 = L(0)R(90)R(0)L(270)L(0)I(90)L(90)I(180)R(180)R(90)I(0)I(0)R(0)I(270)I(270)L(270)      corner = (2, 1) 
#> R | R_1 = I(0)R(0)R(270)L(180)L(270)I(0)L(0)I(90)R(90)R(0)I(270)I(270)R(270)I(180)I(180)L(180)  corner = (1, 2) 
#>     R_2 = L(0)R(90)R(0)L(270)L(0)I(90)L(90)I(180)R(180)R(90)I(0)I(0)R(0)I(270)I(270)I(270)      corner = (2, 1) 
#> L | L_1 = R(0)L(270)L(0)R(90)R(0)I(270)R(270)I(180)L(180)L(270)I(0)I(0)L(0)I(90)I(90)I(90)      corner = (1, 2) 
#>     L_2 = I(0)L(0)L(90)R(180)R(90)I(0)R(0)I(270)L(270)L(0)I(90)I(90)L(90)I(180)I(180)R(180)     corner = (2, 1)