Expand the curve to the next level

Expand the curve to the next level

# S4 method for sfc_hilbert
sfc_expand(p, code)

# S4 method for sfc_peano
sfc_expand(p, code = 1)

# S4 method for sfc_meander
sfc_expand(p, code)

# S4 method for sfc_rules
sfc_expand(p, letters, code = 1L, by = "Cpp")

Arguments

p

An sfc_rules object.

code

The transverse code.

letters

A list patterns, must be a factor.

by

Which implemnetation, only for the testing purpose.

Value

In the same class as the input object. sfc_expand() on the sfc_rules objects returns an sfc_sequence object.

Details

For the Hilbert curve and Meander curve, as long as the transverse code of the first base pattern in the sequence is determinted, the transverse codes for other base patterns in the sequence are all determined. For the Peano curve, code is always 1.

These functions are mainly used internally.

Examples

p = sfc_hilbert("I", 11)
sfc_expand(p, 1)
#> An sfc_hilbert object.
#>   Increase mode: 2 x 2
#>   Expansion rule: Hilbert 
#> 
#> A sequence of 64 base patterns.
#>   R(0)L(270)L(0)R(90)  I(0)R(0)R(270)L(180)
#>   L(270)R(0)R(270)I(180)  R(180)L(90)L(180)I(270)
#>   .... other 4 lines ....
#>   I(90)L(90)L(180)R(270)  I(180)R(180)R(90)L(0)
#>   L(90)R(180)R(90)I(0)  R(0)L(270)L(0)R(90)
p = sfc_peano("I", 11)
sfc_expand(p, 1)
#> An sfc_peano object.
#>   Increase mode: 3 x 3
#>   Expansion rule: Peano 
#> 
#> A sequence of 729 base patterns.
#>   I(0)J(0)R(0)R(270)  I(180)L(180)L(270)J(0)
#>   I(0)J(0)I(0)L(0)  L(90)J(180)R(180)R(90)
#>   .... other 88 lines ....
#>   I(0)J(0)R(0)R(270)  I(180)L(180)L(270)J(0)
#>   I(0)
p = sfc_meander("I", 11)
sfc_expand(p, 1)
#> An sfc_meander object.
#>   Increase mode: 3 x 3
#>   Expansion rule: Meander 
#> 
#> A sequence of 729 base patterns.
#>   R(0)I(270)L(270)I(0)  L(0)L(90)R(180)R(90)
#>   I(0)R(0)I(270)L(270)  I(0)L(0)L(90)R(180)
#>   .... other 88 lines ....
#>   R(0)I(270)L(270)I(0)  L(0)L(90)R(180)R(90)
#>   I(0)
sfc_expand(SFC_RULES_HILBERT, 
    factor(c("I", "R", "L"), levels = sfc_universe(SFC_RULES_HILBERT)))
#> A sequence of 12 base patterns.
#>   R(0)L(270)L(0)R(90)  I(0)R(0)R(270)L(180)
#>   R(0)L(270)L(0)I(90)