Constructor of the sfc_sequence class

Constructor of the sfc_sequence class

sfc_sequence(seq, rot = 0L, universe = NULL)

sfc_seed(seq, rot = 0L, universe = NULL)

sfc_unit(seq, rot = 0L, universe = NULL)

Arguments

seq

A sequence of base patterns. The value can be a vector of letters or a string.

rot

The corresponding rotations of base patterns. If it has length one and the sequence contains R/L/I (right/left/forward), rot controls the rotation of the first base pattern and the rotations for remaining base patterns in the sequence are automatically identified.

universe

The universe of base patterns. A vector of letters.

Details

This funtion is very low-level. Normally, users do not need to directly use this constructor.

sfc_seed class is the same as the sfc_sequence class. It is used specifically as the "seed sequence" when generating the curves.

sfc_unit class is also the same as the sfc_sequence class. It is used specifically when defining the expansion rules.

Examples

sfc_sequence("ABCD", rot = c(0, 90, 180, 270), universe = c("A", "B", "C", "D"))
#> A sequence of 4 base patterns.
#>   A(0)B(90)C(180)D(270)